moved UpdateGpxCategoryTask into separate class
