+++ OsmAnd/src/net/osmand/plus/myplaces/EditTrackGroupDialogFragment.java
@@ -3,12 +3,10 @@ package net.osmand.plus.myplaces;
 import static net.osmand.plus.settings.bottomsheets.BooleanPreferenceBottomSheet.getCustomButtonView;
 import static net.osmand.plus.settings.bottomsheets.BooleanPreferenceBottomSheet.updateCustomButtonView;
 
-import android.app.ProgressDialog;
 import android.content.Context;
 import android.graphics.Typeface;
 import android.os.AsyncTask;
 import android.os.Bundle;
-import android.text.TextUtils;
 import android.view.ContextThemeWrapper;
 import android.view.Gravity;
 import android.view.View;
@@ -22,9 +20,7 @@ import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentActivity;
 import androidx.fragment.app.FragmentManager;
 
-import net.osmand.GPXUtilities;
 import net.osmand.GPXUtilities.GPXFile;
-import net.osmand.GPXUtilities.WptPt;
 import net.osmand.plus.OsmandApplication;
 import net.osmand.plus.R;
 import net.osmand.plus.activities.MapActivity;
@@ -50,15 +46,11 @@ import net.osmand.plus.track.helpers.GpxSelectionHelper.GpxDisplayGroup;
 import net.osmand.plus.track.helpers.GpxSelectionHelper.GpxDisplayItem;
 import net.osmand.plus.track.helpers.GpxSelectionHelper.GpxDisplayItemType;
 import net.osmand.plus.track.helpers.GpxSelectionHelper.SelectedGpxFile;
-import net.osmand.plus.track.helpers.SavingTrackHelper;
 import net.osmand.plus.utils.AndroidUtils;
 import net.osmand.plus.utils.UiUtilities;
 import net.osmand.util.Algorithms;
 
-import java.io.File;
-import java.lang.ref.WeakReference;
 import java.util.HashSet;
-import java.util.List;
 import java.util.Set;
 
 public class EditTrackGroupDialogFragment extends MenuBottomSheetDialogFragment implements OnPointsDeleteListener, OnGroupNameChangeListener {
@@ -159,7 +151,7 @@ public class EditTrackGroupDialogFragment extends MenuBottomSheetDialogFragment 
 						FragmentActivity activity = getActivity();
 						if (activity != null) {
 							FragmentManager fragmentManager = activity.getSupportFragmentManager();
-							RenameTrackGroupBottomSheet.showInstance(fragmentManager, EditTrackGroupDialogFragment.this, group, EditTrackGroupDialogFragment.this);
+							RenameTrackGroupBottomSheet.showInstance(fragmentManager, EditTrackGroupDialogFragment.this, group);
 						}
 					}
 				})
@@ -228,7 +220,7 @@ public class EditTrackGroupDialogFragment extends MenuBottomSheetDialogFragment 
 						FragmentActivity activity = getActivity();
 						if (activity != null) {
 							FragmentManager fragmentManager = activity.getSupportFragmentManager();
-							CopyTrackGroupToFavoritesBottomSheet.showInstance(fragmentManager, EditTrackGroupDialogFragment.this, group, EditTrackGroupDialogFragment.this);
+							CopyTrackGroupToFavoritesBottomSheet.showInstance(fragmentManager, EditTrackGroupDialogFragment.this, group);
 						}
 					}
 				})
@@ -325,7 +317,7 @@ public class EditTrackGroupDialogFragment extends MenuBottomSheetDialogFragment 
 		}
 	}
 
-	private static String getCategoryName(@NonNull Context ctx, String category) {
+	public static String getCategoryName(@NonNull Context ctx, String category) {
 		return Algorithms.isEmpty(category) ? ctx.getString(R.string.shared_string_waypoints) : category;
 	}
 
@@ -363,123 +355,7 @@ public class EditTrackGroupDialogFragment extends MenuBottomSheetDialogFragment 
 	}
 
 	@Override
-	public void onGroupNameChanged() {
+	public void onTrackGroupChanged() {
 		dismiss();
-	}
-
-	public static class UpdateGpxCategoryTask extends AsyncTask<Void, Void, Void> {
-
-		private OsmandApplication app;
-		private WeakReference<FragmentActivity> activityRef;
-
-		private GpxDisplayGroup group;
-
-		private String newCategory;
-		private Integer newColor;
-
-		private ProgressDialog progressDialog;
-		private boolean wasUpdated = false;
-		private UpdateGpxCategoryTask(@NonNull FragmentActivity activity, @NonNull GpxDisplayGroup group) {
-			this.app = (OsmandApplication) activity.getApplication();
-			activityRef = new WeakReference<>(activity);
-
-			this.group = group;
-		}
-
-		public UpdateGpxCategoryTask(@NonNull FragmentActivity activity, @NonNull GpxDisplayGroup group,
-		                             @NonNull String newCategory) {
-			this(activity, group);
-			this.newCategory = newCategory;
-		}
-
-		UpdateGpxCategoryTask(@NonNull FragmentActivity activity, @NonNull GpxDisplayGroup group,
-		                      @NonNull Integer newColor) {
-			this(activity, group);
-			this.newColor = newColor;
-		}
-
-		@Override
-		protected void onPreExecute() {
-			FragmentActivity activity = activityRef.get();
-			if (activity != null) {
-				progressDialog = new ProgressDialog(activity);
-				progressDialog.setTitle(EditTrackGroupDialogFragment.getCategoryName(app, group.getName()));
-				progressDialog.setMessage(newCategory != null ? "Changing name" : "Changing color");
-				progressDialog.setCancelable(false);
-				progressDialog.show();
-
-				GPXFile gpxFile = group.getGpx();
-				if (gpxFile != null) {
-					SavingTrackHelper savingTrackHelper = app.getSavingTrackHelper();
-					List<GpxDisplayItem> items = group.getModifiableList();
-					String prevCategory = group.getName();
-					boolean emptyCategory = TextUtils.isEmpty(prevCategory);
-					for (GpxDisplayItem item : items) {
-						WptPt wpt = item.locationStart;
-						if (wpt != null) {
-							boolean update = false;
-							if (emptyCategory) {
-								if (TextUtils.isEmpty(wpt.category)) {
-									update = true;
-								}
-							} else if (prevCategory.equals(wpt.category)) {
-								update = true;
-							}
-							if (update) {
-								wasUpdated = true;
-								String category = newCategory != null ? newCategory : wpt.category;
-								int color = newColor != null ? newColor : wpt.colourARGB;
-								if (gpxFile.showCurrentTrack) {
-									savingTrackHelper.updatePointData(wpt, wpt.getLatitude(), wpt.getLongitude(),
-											System.currentTimeMillis(), wpt.desc, wpt.name, category, color);
-								} else {
-									gpxFile.updateWptPt(wpt, wpt.getLatitude(), wpt.getLongitude(),
-											System.currentTimeMillis(), wpt.desc, wpt.name, category, color,
-											wpt.getIconName(), wpt.getBackgroundType());
-								}
-							}
-						}
-					}
-				}
-			}
-		}
-
-		@Override
-		protected Void doInBackground(Void... voids) {
-			GPXFile gpxFile = group.getGpx();
-			if (gpxFile != null && !gpxFile.showCurrentTrack && wasUpdated) {
-				GPXUtilities.writeGpxFile(new File(gpxFile.path), gpxFile);
-			}
-			return null;
-		}
-
-		@Override
-		protected void onPostExecute(Void aVoid) {
-			GPXFile gpxFile = group.getGpx();
-			if (gpxFile != null && wasUpdated) {
-				syncGpx(gpxFile);
-			}
-
-			if (progressDialog != null && progressDialog.isShowing()) {
-				progressDialog.dismiss();
-			}
-
-			FragmentActivity activity = activityRef.get();
-			if (activity instanceof MapActivity) {
-				MapActivity mapActivity = (MapActivity) activity;
-				TrackMenuFragment fragment = mapActivity.getTrackMenuFragment();
-				if (fragment != null) {
-					fragment.updateContent();
-				}
-			}
-		}
-
-		private void syncGpx(GPXFile gpxFile) {
-			MapMarkersHelper markersHelper = app.getMapMarkersHelper();
-			MapMarkersGroup group = markersHelper.getMarkersGroup(gpxFile);
-			if (group != null) {
-				markersHelper.runSynchronization(group);
-			}
-		}
 	}
 }
